feat: build-plugin-speed
