+++ packages/icejs/package.json
@@ -36,6 +36,7 @@
     "build-plugin-ice-store": "2.0.6",
     "build-plugin-react-app": "2.1.1",
     "build-plugin-pwa": "1.1.0",
+    "build-plugin-speed": "1.0.0",
     "chalk": "^4.1.0",
     "chokidar": "^3.3.1",
     "commander": "^5.0.0",
