+++ packages/plugin-speed/README.md
@@ -1,0 +1,3 @@
+## build-plugin-speed
+
+> build-scripts plugin for logout build speed
